Changed triangle vertex precision from float to double in multiple subsystems to fix rare problems; Updated FocalEngine; Finalized new multithreading in the MeasurementGrid class; .RUG files now contain double-precision representation of vertices.

+++ SubSystems/ComplexityCore/ComplexityMetricInfo.h
@@ -118,7 +118,7 @@ public:
 // For purposes of complexity metric storing of all of raw data is redundant,but it is needed for saving RUG file.
 struct RawMeshData
 {
-	std::vector<float> Vertices;
+	std::vector<double> Vertices;
 	std::vector<float> Colors;
 	std::vector<float> UVs;
 	std::vector<float> Tangents;
@@ -148,15 +148,13 @@ public:
 	std::vector<int> TriangleSelected;
 
 	RawMeshData MeshData;
-
-	std::vector<std::vector<glm::vec3>> Triangles;
+	std::vector<std::vector<glm::dvec3>> Triangles;
 	std::vector<double> TrianglesArea;
 		
 	std::vector<std::vector<glm::vec3>> TrianglesNormals;
-	std::vector<glm::vec3> TrianglesCentroids;
+	std::vector<glm::dvec3> TrianglesCentroids;
 
-	void FillTrianglesData(std::vector<float>& Vertices, std::vector<float>& Colors, std::vector<float>& UVs, std::vector<float>& Tangents, std::vector<int>& Indices, std::vector<float>& Normals);
-
+	void FillTrianglesData(std::vector<double>& Vertices, std::vector<float>& Colors, std::vector<float>& UVs, std::vector<float>& Tangents, std::vector<int>& Indices, std::vector<float>& Normals);
 		
 	glm::vec3 GetAverageNormal();
 	void UpdateAverageNormal();
